requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
structlog==23.2.0
prometheus-client==0.19.0
httpx==0.25.2
tenacity==8.2.3

# Add missing dependencies
pydantic-settings==2.1.0  # Missing from requirements
psycopg2-binary==2.9.9   # For PostgreSQL support
